+++ src/blog-post/load-blog-post.spec.js
@@ -1,5 +1,5 @@
 import {describe, it} from '../test.js';
-import assert from 'assert';
+import {strict as assert} from 'assert';
 import hamjest from 'hamjest';
 import {loadBlogPost, loadManyBlogPosts} from './load-blog-post.js';
 import {BlogPost} from './BlogPost.js';